refactor(shared, agent/primary): chat protocol schemas + Primary chat_handler 패키지화 (#75 PR 4 commit 15)
(1) shared/chat_protocol/schemas — 도메인별 파일 분리
- 단일 schemas.py → 서브패키지로 분할:
  · `schemas/session.py` — SessionCreateRequest / SessionRead / SessionUpdateRequest
  · `schemas/chat.py` — ChatSendRequest / ChatSendResponse
  · `schemas/events.py` — ChatEvent / ChatEventType
  · `schemas/__init__.py` — re-export (caller import 경로 호환)
- 각 도메인이 독립 파일 — schema 변경 시 다른 도메인 안 건드림 (SRP)

(2) agent/primary chat_handler — 단일 파일 → 패키지
- 단일 chat_handler.py → 패키지로 분할:
  · `chat_handler/session.py` — SessionRuntime + SessionRegistry (in-memory state)
  · `chat_handler/router.py` — make_chat_router (FastAPI route + SSE iter helpers)
  · `chat_handler/worker.py` — run_session_turn (background graph 호출)
  · `chat_handler/__init__.py` — re-export
- 각 모듈이 단일 책임:
  · session: state 자료구조 + lifecycle
  · router: HTTP/SSE 경계 — runtime 가져오고 task spawn / event yield 만
  · worker: graph 호출 phase 별 — _build_graph_input, _stream_chunks_to_session,
    _finalize_agent_response, _publish_agent_chat, _emit_error 로 분해

(3) _run_graph_and_stream → run_session_turn 리팩터링
- 옛 모놀리식 함수 (graph 호출 + 필터 + 누적 + publish + SSE emit 일체) 를
  phase 별 helper 로 분해:
  · `_build_graph_input(session_id, text)` — astream input dict (messages +
    extra_system_message)
  · `_stream_chunks_to_session(graph, runtime, text, message_id)` — graph
    .astream 소비 + classify_response 필터 + chunk push + 누적 텍스트 반환
  · `_finalize_agent_response(runtime, user_message_id, agent_text, event_bus)`
    — publish + SSE message/done
  · `_publish_agent_chat`, `_emit_error` — 기존과 동일 fire-and-forget helper
- `run_session_turn` 본문은 timeout / error 처리 + phase 호출만 — 추상화 일관

검증:
- 단위: shared 68/68 / primary 10/10 pass
- e2e (refactored chat handler 동작): POST /api/sessions → POST /api/chat
  → SSE chunks / done → chats row (user + agent) 정합 ✅
